+++ controlgallery/AtomUIGallery/ShowCases/DataEntry/RadioButton/ViewModels/RadioButtonViewModel.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Globalization;
 using System.Reactive;
 using AtomUI.Controls;
 using AtomUI.Data;
@@ -41,6 +42,41 @@ public class RadioButtonViewModel : ReactiveObject, IRoutableViewModel
         set => this.RaiseAndSetIfChanged(ref _radioOptions, value);
     }
 
+    private IList<RadioButtonOption>? _twoWayRadioOptions;
+
+    public IList<RadioButtonOption>? TwoWayRadioOptions
+    {
+        get => _twoWayRadioOptions;
+        set => this.RaiseAndSetIfChanged(ref _twoWayRadioOptions, value);
+    }
+
+    private object? _twoWayCheckedItem;
+
+    public object? TwoWayCheckedItem
+    {
+        get => _twoWayCheckedItem;
+        set
+        {
+            if (ReferenceEquals(_twoWayCheckedItem, value))
+            {
+                return;
+            }
+
+            this.RaiseAndSetIfChanged(ref _twoWayCheckedItem, value);
+            UpdateTwoWayCheckedSummary();
+        }
+    }
+
+    private string? _twoWayCheckedSummary;
+
+    public string? TwoWayCheckedSummary
+    {
+        get => _twoWayCheckedSummary;
+        set => this.RaiseAndSetIfChanged(ref _twoWayCheckedSummary, value);
+    }
+
+    private RadioButtonOption? _twoWayChengduOption;
+
     private bool _toggleDisabledRadioUnCheckedEnabled;
 
     public bool ToggleDisabledRadioUnCheckedEnabled
@@ -58,6 +94,8 @@ public class RadioButtonViewModel : ReactiveObject, IRoutableViewModel
     }
 
     public ReactiveCommand<Unit, Unit> ToggleDisabledCommand { get; }
+    public ReactiveCommand<Unit, Unit> SelectChengduCommand { get; }
+    public ReactiveCommand<Unit, Unit> ClearTwoWayCheckedItemCommand { get; }
 
     public RadioButtonViewModel(IScreen screen)
     {
@@ -67,12 +105,65 @@ public class RadioButtonViewModel : ReactiveObject, IRoutableViewModel
         _toggleDisabledRadioCheckedEnabled = true;
 
         ToggleDisabledCommand = ReactiveCommand.Create(HandleToggleDisabled);
+        SelectChengduCommand  = ReactiveCommand.Create(HandleSelectChengdu);
+        ClearTwoWayCheckedItemCommand = ReactiveCommand.Create(HandleClearTwoWayCheckedItem);
     }
 
     private void HandleToggleDisabled()
     {
         ToggleDisabledRadioUnCheckedEnabled = !ToggleDisabledRadioUnCheckedEnabled;
         ToggleDisabledRadioCheckedEnabled = !ToggleDisabledRadioCheckedEnabled;
+    }
+
+    public void ConfigureTwoWayRadioOptions(
+        RadioButtonOption hangzhou,
+        RadioButtonOption shanghai,
+        RadioButtonOption beijing,
+        RadioButtonOption chengdu)
+    {
+        _twoWayChengduOption = chengdu;
+        TwoWayRadioOptions =
+        [
+            hangzhou,
+            shanghai,
+            beijing,
+            chengdu
+        ];
+        TwoWayCheckedItem = shanghai;
+    }
+
+    public void ClearTwoWayRadioOptions()
+    {
+        _twoWayChengduOption = null;
+        TwoWayCheckedItem    = null;
+        TwoWayRadioOptions   = null;
+        TwoWayCheckedSummary = null;
+    }
+
+    private void HandleSelectChengdu()
+    {
+        if (_twoWayChengduOption != null)
+        {
+            TwoWayCheckedItem = _twoWayChengduOption;
+        }
+    }
+
+    private void HandleClearTwoWayCheckedItem()
+    {
+        TwoWayCheckedItem = null;
+    }
+
+    private void UpdateTwoWayCheckedSummary()
+    {
+        var selectedText = RadioButtonShowCaseLanguage.Get(RadioButtonShowCaseLangResourceKind.P2ContentNone, "None");
+        if (TwoWayCheckedItem is RadioButtonOption { Content: not null } option)
+        {
+            selectedText = option.Content.ToString() ?? selectedText;
+        }
+
+        TwoWayCheckedSummary = string.Format(CultureInfo.CurrentCulture,
+            RadioButtonShowCaseLanguage.Get(RadioButtonShowCaseLangResourceKind.P2CheckedItemSummaryFormat, "Selected: {0}"),
+            selectedText);
     }
 
     public void EnsureApiRows()